% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_geno_by_fam.R
\name{aggregate.geno.by.fam}
\alias{aggregate.geno.by.fam}
\title{Aggregate genotypes among affected individuals for each family}
\usage{
\method{aggregate}{geno.by.fam}(pedfile)
}
\arguments{
\item{pedfile}{is a genotype file in ped format: A .ped file}
}
\value{
A list with the ped file corrected and aggregated by family and index each variants observed in families
}
\description{
This function aggregates variants across family members. It takes a pedfile as first parameter.
The six first columns must be family id, individual id, father and mother ids, sex and phenotype
Homozyguous variants are replaced by their heterozyguous configurations.
}
