% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootRetroFunRVS.R
\name{bootRetroFunRVS}
\alias{bootRetroFunRVS}
\title{Function computing bootstrap scores}
\usage{
bootRetroFunRVS(
  agg.genos.by.fam,
  n.unique.config.by.fam,
  Z_annot = NULL,
  W = NULL,
  prob.sharing.by.fam,
  nullval,
  nrep = 1000
)
}
\arguments{
\item{agg.genos.by.fam}{a list return by agg.genos.by.fam}

\item{n.unique.config.by.fam}{a list with the corresponding genotype configurations in each family}

\item{Z_annot}{a p*q matrix of functional annotations. The first column should be composed only with ones}

\item{W}{a vector of weights. Should be a unitary vector when unweighted version is needed.}

\item{prob.sharing.by.fam}{a list with the corresponding sharing probabilities for each genotype configuration}

\item{nullval}{a dataframe with four colums (FamID, Expected, Variance and Covariance) return by compute.null function}

\item{nrep}{the number of bootstrap samples to generate}
}
\value{
a matrix of scores where each column is an annotation and each row is a bootstrap sample
}
\description{
This function computes the burden score on non-parametric bootstrap samples. Genotypes are resampled based on
sharing probabilities
}
