% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Burden_by_Annot.R
\name{compute.Burden.by.Annot}
\alias{compute.Burden.by.Annot}
\title{Compute the Burden Statistic for q functional annotations}
\usage{
compute.Burden.by.Annot(null.value.by.fam, aggregate.geno.by.fam, Z_annot, W)
}
\arguments{
\item{null.value.by.fam}{is a dataframe with four colums (FamID, Expected, Variance and Covariance) return by compute.null function}

\item{aggregate.geno.by.fam}{is the list returned by aggregate.geno.by.fam function}

\item{Z_annot}{is a p*q matrix of functional annotations. The first column should be composed only with ones}

\item{W}{is a p*p matrix of weights. Should be a identity matrix when unweighted version is needed.}
}
\value{
a list with each Score by annotation
}
\description{
This function computes the functional annotation burden statistic. It assumes that the first parameter is
the data.frame with null values for each family with the corresponding variance and covariance. The second parameter should be
a list with aggregated genotypes by family with related variant index. Statistics are computed based on Z_annot which is a matrix where first column is only ones.
Weights are allowed through the W parameter
}
