% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_expected_by_fam.R
\name{compute.expected.by.fam}
\alias{compute.expected.by.fam}
\title{Function computing the genotype expected value for each family}
\usage{
compute.expected.by.fam(
  configs.with.probs,
  distinguishHomo = FALSE,
  cryptic.relatedness = FALSE
)
}
\arguments{
\item{configs.with.probs}{a list with configurations and corresponding probabilities by family}

\item{distinguishHomo}{a boolean: TRUE when inbreeding among family members is expected, FALSE otherwise
#' @param cryptic.relatedness a boolean: TRUE when cryptic relatedness is expected, FALSE otherwise}
}
\value{
a dataframe with the genotype expected value for each family
}
\description{
Function computing the genotype expected value for each family
}
