% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_null_by_indiv.r
\name{compute.null.by.indiv}
\alias{compute.null.by.indiv}
\title{Compute the null value for pedigrees}
\usage{
compute.null.by.indiv(
  pedigree,
  distinguishHomo = FALSE,
  cryptic.relatedness = FALSE,
  kinshipCoeff = NULL,
  eta = 0.95,
  etap = 0.95,
  probs.list = NULL
)
}
\arguments{
\item{pedigree}{is a object of pedigree format}

\item{eta}{multiplicative factor for the joint carrier probability for computing the covariance between two variants for a pairs of subjects (between 0 and 1)}

\item{etap}{multiplicative factor for the marginal carrier probability for computing the covariance between two variants in the same subject (between 0 and 1)}
}
\value{
List over pedigrees of lists of the individual variance-covariance matrix for one variant and the individual covariance matrix for two variants

List of sharing configuration probabilities computed by the RVsharing function from the RVS package
}
\description{
This function computes the genotype null value for one or more families. It assumes that the
the first parameter is a list with all the possible family configurations while the second
parameter is a list with all the sharing probabilities for the all family configurations
}
