% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_null_parallel.R
\name{compute.null.parallel}
\alias{compute.null.parallel}
\title{Compute the null value for pedigrees using parallel programming}
\usage{
compute.null.parallel(
  pedigree,
  distinguishHomo = FALSE,
  cryptic.relatedness = FALSE,
  kinshipCoeff = NULL,
  fam.ids = NULL,
  out.prob.dist = FALSE
)
}
\arguments{
\item{pedigree}{is a object of pedigree format}

\item{distinguishHomo}{a boolean: TRUE when inbreeding among family members is expected, FALSE otherwise}

\item{cryptic.relatedness}{a boolean: TRUE when cryptic relatedness is expected, FALSE otherwise}

\item{kinshipCoeff}{mean kinship coefficient among the pedigree founders}

\item{fam.ids}{a vector of family ids where inbreeding or cryptic relatedness may be computed}

\item{out.prob.dist}{a boolean: TRUE when the probability distributions need to be a part of the output}
}
\value{
The expected genotype value, variance and covariance for each pedigree within a data.frame
}
\description{
This function computes the genotype null value for one or more families. It assumes that the
the first parameter is a list with all the possible family configurations while the second
parameter is a list with all the sharing probabilities for the all family configurations
}
