% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_null.R
\name{compute.null}
\alias{compute.null}
\title{Compute the null value for pedigrees}
\usage{
compute.null(
  pedigree,
  distinguishHomo = FALSE,
  cryptic.relatedness = FALSE,
  kinshipCoeff = NULL,
  fam.ids = NULL
)
}
\arguments{
\item{pedigree}{is a object of pedigree format}

\item{distinguishHomo}{a boolean: TRUE when inbreeding among family members is expected, FALSE otherwise}

\item{cryptic.relatedness}{a boolean: TRUE when cryptic relatedness is expected, FALSE otherwise}

\item{kinshipCoeff}{mean kinship coefficient among the pedigree founders}

\item{fam}{a vector of family ids where inbreeding or cryptic relatedness may be computed}
}
\value{
The expected genotype value, variance and covariance for each pedigree within a data.frame
}
\description{
This function computes the genotype null value for one or more families. It assumes that the
the first parameter is a list with all the possible family configurations while the second
parameter is a list with all the sharing probabilities for the all family configurations
}
