% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Var_by_Annot.R
\name{compute.Var.by.Annot}
\alias{compute.Var.by.Annot}
\title{Compute the Variance for q functional annotations
This function computes the functional annotation variance statistic. It assumes that the first parameter is
the data.frame with null values for each family with the corresponding variance and covariance. The second parameter should be
a list with aggregated genotypes by family with related variant index. Statistics are computed based on Z_annot which is a matrix where first column is only ones.
Weights are allowed through the W parameter. Haplotype structure can be assumed with the independence parameter.}
\usage{
compute.Var.by.Annot(
  null.value.by.fam,
  aggregate.geno.by.fam,
  Z_annot,
  W,
  independence = F
)
}
\arguments{
\item{null.value.by.fam}{is a dataframe with four colums (FamID, Expected, Variance and Covariance) return by compute.null function}

\item{aggregate.geno.by.fam}{is the list returned by aggregate.geno.by.fam function}

\item{Z_annot}{is a p*q matrix of functional annotations. The first column should be composed only with ones}

\item{W}{is a vector of weights. Should be a unitary vector when unweighted version is needed.}

\item{independence}{is a boolean. Default value is FALSE. If variant independence can be assumed, use independence=TRUE.}
}
\value{
a list with each variance by annotation
}
\description{
Compute the Variance for q functional annotations
This function computes the functional annotation variance statistic. It assumes that the first parameter is
the data.frame with null values for each family with the corresponding variance and covariance. The second parameter should be
a list with aggregated genotypes by family with related variant index. Statistics are computed based on Z_annot which is a matrix where first column is only ones.
Weights are allowed through the W parameter. Haplotype structure can be assumed with the independence parameter.
}
