% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_var_by_fam_parallel.R
\name{compute.var.by.fam.parallel}
\alias{compute.var.by.fam.parallel}
\title{Function computing the theoretical variance and covariance for each family}
\usage{
compute.var.by.fam.parallel(
  configs.with.probs,
  distinguishHomo = FALSE,
  cryptic.relatedness = FALSE
)
}
\arguments{
\item{configs.with.probs}{a list with configurations and corresponding probabilities by family}

\item{distinguishHomo}{a boolean: TRUE when inbreeding among family members is expected, FALSE otherwise
#' @param cryptic.relatedness a boolean: TRUE when cryptic relatedness is expected, FALSE otherwise}
}
\value{
a dataframe with the genotype variance for each family
}
\description{
Function computing the theoretical variance and covariance for each family
}
