% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_var_by_fam.R
\name{compute.var.by.fam}
\alias{compute.var.by.fam}
\title{Function computing the theoretical variance and covariance for each family}
\usage{
compute.var.by.fam(config.by.fam, sharing.proba.by.fam)
}
\arguments{
\item{config.by.fam}{a list with configuration by family}

\item{sharing.proba.by.fam}{a list with sharing probabilities by family}
}
\value{
a dataframe with Famid, variance and covariance
}
\description{
Function computing the theoretical variance and covariance for each family
}
