% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_genos_by_fam.R
\name{resample.genos.by.fam}
\alias{resample.genos.by.fam}
\title{Function resampling genotypes by fam}
\usage{
resample.genos.by.fam(aggregate.geno.by.fam, prob_sharing_by_famid = NULL)
}
\arguments{
\item{aggregate.geno.by.fam}{a list return by aggregate.geno.by.fam}

\item{prob_sharing_by_famid}{a list with the corresponding sharing probabilities for each genotype configurations in families}
}
\value{
resample data.frame
}
\description{
This functions proceeds to non-parametric bootstrap for each family. Genotypes can be resampled based on
sharing probabilities
}
