% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetroFun_RVS.R
\name{RetroFun.RVS}
\alias{RetroFun.RVS}
\title{Compute the p-values associated with each functional annotation}
\usage{
RetroFun.RVS(
  null.value.by.fam,
  aggregate.geno.by.fam,
  Z_annot,
  W,
  independence = F
)
}
\arguments{
\item{null.value.by.fam}{is a dataframe with four colums (FamID, Expected, Variance and Covariance) return by compute.null function}

\item{aggregate.geno.by.fam}{is the list returned by aggregate.geno.by.fam function}

\item{Z_annot}{is a p*q matrix of functional annotations. The first column should be composed only with ones}

\item{W}{is a p*p matrix of weights. Should be a identity matrix when an unweighted approach is needed.}

\item{independence}{is a boolean. Default value is FALSE. If variant independence can be assumed, use independence=TRUE.}
}
\value{
a data.frame of p-values for each score by annotation. ACAT and Fisher combined p-values are computed.
}
\description{
This function computes both the ACAT-combined and Fisher's method p-values
}
