% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_covar_by_fam.R
\name{update.covar.by.fam}
\alias{update.covar.by.fam}
\title{Compute the p-values associated with each functional annotation}
\usage{
update.covar.by.fam(null.value.by.fam, null.covar.by.indiv)
}
\arguments{
\item{null.value.by.fam}{is a dataframe with four colums (FamID, Expected, Variance and Covariance) returned by the compute.null function}

\item{null.covar.by.indiv}{is a list of variances and covariances for each pair of affected subjects in every family returned by the compute.null.by.indiv function}
}
\value{
The expected genotype value, variance and covariance for each pedigree within a data.frame
}
\description{
This function computes both the ACAT-combined and Fisher's method p-values
}
