% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_consanguinity.R
\name{.check.consanguinity}
\alias{.check.consanguinity}
\title{usethis::use_package("kinship2")
Check whether there is consanguinity among pedigrees}
\usage{
.check.consanguinity(pedigrees)
}
\arguments{
\item{pedigrees}{is a pedigree of a pedigreeList format or list of pedigrees}
}
\description{
usethis::use_package("kinship2")
Check whether there is consanguinity among pedigrees
}
