% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_genos_by_fam.R
\name{resample.genos.by.fam}
\alias{resample.genos.by.fam}
\title{Function resampling genotypes by fam}
\usage{
resample.genos.by.fam(
  agg.genos.by.fam,
  n.unique.config.by.fam,
  prob.sharing.by.fam = NULL
)
}
\arguments{
\item{agg.genos.by.fam}{a list return by agg.genos.by.fam}

\item{n.unique.config.by.fam}{a list with the corresponding genotype configurations in each family}

\item{prob.sharing.by.fam}{a list with the corresponding sharing probabilities for each genotype configuration}
}
\value{
resample data.frame
}
\description{
This functions proceeds to non-parametric bootstrap for each family. Genotypes can be resampled based on
sharing probabilities
}
